refactor: split oversized coin capture pages
Extract shared inline camera capture UI and lookup draft normalization so no Vue route page exceeds 1k lines. Preserves the camera-first interaction and user-tapped Start Camera behavior for #370.

Constitution: Principle IV + §17

Closes #370
