+++ src/web/src/__tests__/ui-patterns.test.ts
@@ -63,16 +63,17 @@ describe('UI pattern recipes', () => {
   it('keeps Identify Coin camera-first with Add Coin upload icon pattern', () => {
     const lookupPage = readRepoFile(join('pages', 'CoinLookupPage.vue'))
     const addCoinPage = readRepoFile(join('pages', 'AddCoinPage.vue'))
+    const inlineCameraPanel = readRepoFile(join('components', 'InlineCameraCapturePanel.vue'))
 
-    expect(lookupPage).toContain('ref="cameraVideo"')
-    expect(lookupPage).toContain('Start Camera')
-    expect(lookupPage).toContain('@click="startCamera"')
-    expect(addCoinPage).toContain('Start Camera')
-    expect(addCoinPage).toContain('@click="startCamera"')
+    expect(lookupPage).toContain('InlineCameraCapturePanel')
+    expect(addCoinPage).toContain('InlineCameraCapturePanel')
+    expect(inlineCameraPanel).toContain('ref="cameraVideo"')
+    expect(inlineCameraPanel).toContain('Start Camera')
+    expect(inlineCameraPanel).toContain('@click="startCamera"')
     expect(addCoinPage).not.toContain('await startCamera()')
-    expect(lookupPage).toContain('class="shutter-btn"')
-    expect(lookupPage).toContain('class="upload-icon-btn"')
-    expect(lookupPage).toContain('<Images :size="20" />')
+    expect(inlineCameraPanel).toContain('class="shutter-btn"')
+    expect(inlineCameraPanel).toContain('class="upload-icon-btn"')
+    expect(inlineCameraPanel).toContain('<Images :size="20" />')
     expect(lookupPage).not.toContain('CameraCaptureModal')
     expect(lookupPage).not.toContain('Take Photo')
     expect(lookupPage).not.toContain('Upload Image')
